More intellegent time parsing
Be more careful parsing times submitted as CLI options, defaulting to
the current time and issuing a warning when misformatted options are
passed.

Put some guards in place before passing a string to Time.parse, since it
is over optimistic and will convert "=18 minutes ago" to the 18th of the
current month.