Fix two issues with month command
The month command didn't previously work in December.  It didn't handle
the logic to understand that January (month 1) follows December (month
12)

Also it's not necessary to call Time#to_date in the month command.  This
method is private in ruby 1.8.6.  Removing this call provides some
additional level of compatibility running under 1.8.6.