Update the resume action to allow resuming by id, removes some old functionality
This changes the resume action to allow you to either resume by an id,
or resume the last checked out entry.

This removes the functionality that allowed you to change the note on
the last entry, or start a new entry with the new note. IMO the former
should just be accomplished by using the edit action, and the latter
with the in action. I don't think there's any need to duplicate the
functionality of those actions within the resume action.

Addresses #114