Use Date.today to determine start and end for 't week'
If the day of the week for today's date is not the same as the defined start of
the week, then the most recent start of the week date is used.

The 'week' command used Chronic to determine what date to use for the start of
the week. Chronic used today's date, but this makes it difficult to test. Using
'Date.today' allows tests to set today's date.

The end date argument set to 'today' is also added to aid in testing.