Add edit option to clear note
Closes: https://github.com/samg/timetrap/issues/164

The regex used to see if a note has been provided cannot find empty
strings. It is therefore not possible to look for empty string to clear
a note.

Instead a new flag (-c or --clear) clears the last entry's note or it
can be used with the -i flag to clear another entry's note.