Add rubocop for linting
Includes
 - CI github workflow
 - Rubocop and Standard ruby gems
 - Rubocop config

To automate linting and code standards.

RuboCop is a Ruby static code analyzer (a.k.a. linter) and code formatter.
Out of the box it will enforce many of the guidelines outlined in the
community Ruby Style Guide.

StandardRB is built on top of RuboCop with some specific configurations for RuboCop's cops.

Further reading:
 - https://docs.rubocop.org/rubocop/1.18/index.html
 - https://rubystyle.guide/
 - https://github.com/testdouble/standard
 - https://github.com/testdouble/standard/blob/main/config/base.yml